separate message_builder and message rules
